src/uc_ball_hyp_generator/labelingtool/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uc-ball-hyp-generator-labelingtool"
version = "0.1.0"
description = "Qt-based soccer ball labeling tool"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0,<7",
    "PySide6>=6.6,<7",
    "numpy>=1.24",
    "pandas>=2.0",
    "segment-anything>=1.0.0",
]
authors = [{ name = "UC Ball Team" }]
license = { text = "MIT" }

[project.urls]
Homepage = "https://example.com"

[tool.ruff]
line-length = 120
